Split out model API adapter
