+++ src/easy_agents/core/model/__init__.py
@@ -1,3 +1,4 @@
-from .model import AssistantResponse, Model, ModelTokenLimitExceededError
+from .model import Model
+from .types import AssistantResponse, ModelTokenLimitExceededError
 
 __all__ = ["Model", "ModelTokenLimitExceededError", "AssistantResponse"]
