+++ src/easy_agents/core/model/api_adapter/__init__.py
@@ -1,0 +1,15 @@
+from .api_adapter import ModelApiAdapter
+from .basic_adapters import (
+    ApiAdapterExtractXmlReasoningFromContent,
+    ApiAdapterSetAutoToolsAsNone,
+    ApiAdapterSetRequiredToolsAsPartOfSystemPrompt,
+    ApiAdapterStructuredOutputAsTool,
+)
+
+__all__ = [
+    "ModelApiAdapter",
+    "ApiAdapterStructuredOutputAsTool",
+    "ApiAdapterExtractXmlReasoningFromContent",
+    "ApiAdapterSetRequiredToolsAsPartOfSystemPrompt",
+    "ApiAdapterSetAutoToolsAsNone",
+]
